HERO

+++ Programacion/ShutterEarth/src/shutterearth/map/LiveDisplayer.java
@@ -83,7 +83,7 @@ public class LiveDisplayer extends Scene
     }
 
     @Override
-    public void Update(GameContainer gc, int t) throws SlickException
+    public void Update(GameContainer gc, float t) throws SlickException
     {
         if (displayingHealth < actualHealth)
         {
