+++ Programacion/ShutterEarth/src/shutterearth/map/TextDisplayer.java
@@ -63,7 +63,7 @@ public class TextDisplayer extends Scene implements InputProviderListener
     }
 
     @Override
-    public void Update(GameContainer gc, int t) throws SlickException
+    public void Update(GameContainer gc, float t) throws SlickException
     {
         switch (step)
         {
