+++ Programacion/ShutterEarth/src/shutterearth/screens/Store.java
@@ -153,7 +153,7 @@ public class Store extends Scene implements InputProviderListener
     }
 
     @Override
-    public void Update(GameContainer gc, int t) throws SlickException
+    public void Update(GameContainer gc, float t) throws SlickException
     {
         if (clicked)
         {
